outsource staysInStation dataprovider to its own provider
